backend/supervisor-agent/pyproject.toml
[tool.poetry]
name = "supervisor-agent"
version = "0.1.0"
description = "Supervisor Agent with A2A protocol support"
authors = ["Developer <dev@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
boto3 = "^1.40.0"
httpx = "^0.27.0"
pyyaml = "^6.0"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
jsq-ai-sdk = {path = "/Users/yashac/code/jsq-ai-sdk-python", develop = true}

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.0"
mypy = "^1.16.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py311"
